Adding auth.js to protect routes.

+++ src/app/api/feed/[id]/route.ts
@@ -1,16 +1,20 @@
 import { PrismaClient } from '@prisma/client'
 import { NextResponse } from "next/server";
+import { auth } from '@/auth';
 
 const prisma = new PrismaClient();
 
 type Props = {
-  params: Promise<{
-    id: string
-  }>
+  params?: Record<string, string | string[]>
 }
 
-export async function GET(request: Request, props: Props ) {
-  const id = (await props.params).id
+export const GET = auth(async(request, props: Props ) => {
+  if (!props.params)
+    return new NextResponse(null, {
+      status: 400,
+    });
+
+  const id = props.params.id as string;
   const post = await prisma.post.findUnique({
     where: {
       id: parseInt(id),
@@ -20,10 +24,15 @@ export async function GET(request: Request, props: Props ) {
   return new NextResponse(JSON.stringify(post), {
     status: 200,
   });
-};
+});
 
 export async function PUT(request: Request, props: Props ) {
-  const id = (await props.params).id
+  if (!props.params)
+    return new NextResponse(null, {
+      status: 400,
+    });
+
+  const id = props.params.id as string;
   const data = await request.json();
 
   await prisma.post.update({
@@ -39,7 +48,12 @@ export async function PUT(request: Request, props: Props ) {
 }
 
 export async function DELETE(request: Request, props: Props) {
-  const id = (await props.params).id
+  if (!props.params)
+    return new NextResponse(null, {
+      status: 400,
+    });
+
+  const id = props.params.id as string;
   await prisma.post.delete({
     where: {
       id: parseInt(id),
